# ARROW-11354: [Rust] Speed-up cast of dates and times (2-4x)