--- a/rust/arrow/src/compute/kernels/mod.rs
+++ b/rust/arrow/src/compute/kernels/mod.rs
@@ -19,6 +19,7 @@
 
 pub mod aggregate;
 pub mod arithmetic;
+pub mod arity;
 pub mod boolean;
 pub mod cast;
 pub mod comparison;
@@ -30,4 +31,3 @@ pub mod sort;
 pub mod substring;
 pub mod take;
 pub mod temporal;
-pub mod unary;