@jorgecarleitao 2021-01-23
NOTE: this PR is 40 LOC change. The rest comes from the other PR. Please wait for the merge, this is only a draft to indicate what can be done with this.